[CL 3236454 by Lauren Ridge in Dev-VREditor branch]

+++ Engine/Source/Runtime/OpenGLDrv/Private/Mac/OpenGLMac.cpp
@@ -756,7 +756,7 @@ bool PlatformGetAvailableResolutions(FScreenResolutionArray& Resolutions, bool b
 	if (AllModes)
 	{
 		const int32 NumModes = CFArrayGetCount(AllModes);
-		const int32 Scale = FMacApplication::GetPrimaryScreenBackingScaleFactor();
+		const int32 Scale = (int32)FMacApplication::GetPrimaryScreenBackingScaleFactor();
 
 		for (int32 Index = 0; Index < NumModes; Index++)
 		{
